requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.1.0
cffi==1.15.1
charset-normalizer==2.1.1
discord.py==2.1.0
flake8==5.0.4
frozenlist==1.3.3
idna==3.4
mccabe==0.7.0
multidict @ file:///C:/Users/Justi/Downloads/multidict-6.0.2-cp311-cp311-win_amd64.whl
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyNaCl==1.5.0
python-dotenv==0.21.0
yarl @ file:///C:/Users/Justi/Downloads/yarl-1.8.1-cp311-cp311-win_amd64.whl
